chore: extract PatchPanel
