+++ demo3PNG/src/screens/PatchPanel.module.css
@@ -1,0 +1,27 @@
+.container {
+  width: 360px;;
+  background-color: #ffffff;
+}
+
+.avatarContainer {
+  width: 360px;
+  height: 360px;
+  display: block;
+  position: relative;
+  background-color: #eeeeee;
+}
+
+.btnContainer {
+  margin-top: 20px;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+
+.btnContainer button {
+  display: block;
+  width: 120px;
+  height: 40px;
+  margin: 10px;
+  cursor: pointer;
+}
